Support for operational data from netconf client.
