+++ lib/puppet/provider/cisco_yang/cisco.rb
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 require_relative '../../../util/node_util' if Puppet.features.node_util?
+require_relative '../../../util/yang_accessor' if Puppet.features.node_util?
 
 Puppet::Type.type(:cisco_yang).provide(:cisco) do
   desc 'IOS-XR configuration management via YANG.'
@@ -22,12 +23,14 @@ Puppet::Type.type(:cisco_yang).provide(:cisco) do
   confine feature: :node_util
 
   def initialize(value={})
+puts "====> initialize - value: #{value}"
     super(value)
     @node = Cisco::Node.instance(Cisco::Client::GRPC)
     debug 'Created provider instance of cisco_yang.'
   end
 
   def exists?
+puts "====> exists?"
     activate
     source && source != :absent
   end
