+++ lib/util/show_running_yang.rb
@@ -18,171 +18,11 @@
 # sources).
 
 require 'optparse'
-require_relative 'node_util'
+require_relative 'yang_accessor'
 
-module Cisco
-  # Utility class to output the current state of an XR configuration.
-  class YangViewer
-    def process(options)
-      @options = options
+# Utility class to output the current state of an XR configuration.
 
-      client # initialize the client
-
-      dir_or_file = options[:path]
-
-      file = nil
-      dir = nil
-
-      if File.exist?(dir_or_file)
-        if File.directory?(dir_or_file)
-          dir = dir_or_file
-        else
-          file = dir_or_file
-        end
-      else
-        puts "Directory or file not found: #{dir_or_file}"
-        exit(-1)
-      end
-
-      puts "File found: #{file}" if file
-      puts "Directory found: #{dir}" if dir
-      puts 'Searching for configuration data...' unless @options[:oper]
-      puts 'Searching for operational data...' if @options[:oper]
-
-      t1 = Time.now
-
-      @files = 0
-      @cnrs = 0
-      @errors = 0
-      if file
-        process_file(file)
-        @files += 1
-      else
-        Dir.glob(dir + '/*.yang').sort.each do |item|
-          process_file(item)
-          @files += 1
-        end
-      end
-
-      delta = Time.now - t1
-      puts '---------------------------------------------'
-      puts "Files Processed: #{@files}"
-      puts "Containers Processed: #{@cnrs}"
-      puts "Errors: #{@errors}"
-      puts "Time: #{delta.round(2)} seconds"
-      puts # spacer
-    end
-
-    def process_file(file)
-      @module = nil
-      @namespace = nil
-      @containers = {}
-      puts "[ Processing file #{file} ]" if @options[:verbose]
-
-      File.open(file) do |f|
-        loop do
-          break if (line = f.gets).nil?
-          process_line(line, f)
-        end
-      end
-    end
-
-    def process_line(line, file)
-      if @module.nil?
-        @module = Regexp.last_match(1) if line =~ /^module (.+) {/
-      elsif @namespace.nil?
-        @namespace = Regexp.last_match(1) if line =~ /^  namespace "(.+)"/
-      elsif line =~ /^  container (.+) {/
-        process_root_container(@module, @namespace, Regexp.last_match(1), file)
-      end
-    end
-
-    def process_root_container(module_name, namespace, container, file)
-      operation = :get_config
-      loop do
-        line = file.gets
-        break if !line || line.strip == ''
-        if line =~ /^    config false;/ # abort cnr if not config
-          operation = :get_oper
-          break
-        end
-      end
-
-      # only output config or operational data, depending on options
-      if @options[:oper]
-        return if operation == :get_config
-      else
-        return if operation != :get_config
-      end
-
-      # guard against duplicate containers
-      if @containers.key?(container)
-        puts "[   Duplicate container #{container} ]" if @options[:verbose]
-        return
-      end
-
-      yang_target = client.yang_target(module_name, namespace, container)
-
-      @containers[container] = true
-      @cnrs += 1
-
-      begin
-        data = client.get(data_format: :yang_json,
-                          command:     yang_target,
-                          mode:        operation)
-        if data && data.strip.length > 0
-          puts "[   Processing container #{container} ]"\
-              if @options[:verbose]
-          output_data(yang_target, data)
-        else
-          puts "[   Processing container #{container}: no data returned ]"\
-              if @options[:verbose]
-        end
-      rescue Cisco::ClientError, Cisco::YangError => e
-        puts "[   Processing container #{container} ]" if @options[:verbose]
-        @errors += 1
-        puts "!!Error on '#{yang_target}': #{e}"
-        debug e.backtrace
-        puts # spacer
-      end
-    end
-
-    def output_data(yang_target, data)
-      if @options[:manifest]
-        if @options[:client_class] == Cisco::Client::GRPC
-          puppet_type = 'cisco_yang'
-          ensure_prop = "    ensure => present,\n"
-        else
-          puppet_type = 'cisco_yang_netconf'
-          ensure_prop = ''
-        end
-
-        puts "  #{puppet_type} { '#{yang_target}':\n#{ensure_prop}"\
-            "    source => '#{data.chomp.gsub(/\n/, "\n    ")}'\n"\
-            '  }'
-      else
-        puts data
-      end
-      puts # spacer
-    end
-
-    def client
-      unless @client
-        @client = Cisco::Client.create(@options[:client_class])
-
-        puts "[ Connected to client: #{@client} ]"\
-            if @options[:verbose]
-      end
-      @client
-    rescue Cisco::AuthenticationFailed
-      abort 'Unauthorized to connect'
-    rescue Cisco::ClientError, TypeError, ArgumentError => e
-      abort "Error in establishing connection: #{e}"
-    end
-  end # YangViewer
-end # Cisco
-
-options = { path: '/pkg/yang', environment: 'default' }
+options = {}
 optparse = OptionParser.new do |opts|
   opts.banner = 'Usage: ruby [path]show_running_yang.rb [options] [file_or_directory_path]'
 
@@ -246,5 +86,5 @@ elsif ARGV.length > 1
   exit(-1)
 end
 
-ycv = Cisco::YangViewer.new
-ycv.process(options)
+ya = Cisco::YangAccessor.new
+ya.process(options)
